some more formatting? idk

+++ src/cpp/code_16.cpp
@@ -22,8 +22,8 @@ public:
         sort(nums.begin(), nums.end());
         int soln;
         for (size_t i = 0; i < nums.size(); i++) {
-            int j = i + 1;
-            int k = nums.size() - 1;
+            size_t j = i + 1;
+            size_t k = nums.size() - 1;
             while (j < k) {
                 int sum = nums[i] + nums[j] + nums[k];
                 if (sum == target) {
